+++ capa/features/extractors/ghidra/file.py
@@ -23,13 +23,27 @@ import capa.features.extractors.helpers
 import capa.features.extractors.strings
 import capa.features.extractors.ghidra.helpers
 from capa.features.file import Export, Import, Section, FunctionName
-from capa.features.common import FORMAT_PE, FORMAT_ELF, Format, String, Feature, Characteristic
-from capa.features.address import NO_ADDRESS, Address, FileOffsetAddress, AbsoluteVirtualAddress
+from capa.features.common import (
+    FORMAT_PE,
+    FORMAT_ELF,
+    Format,
+    String,
+    Feature,
+    Characteristic,
+)
+from capa.features.address import (
+    NO_ADDRESS,
+    Address,
+    FileOffsetAddress,
+    AbsoluteVirtualAddress,
+)
 
 MAX_OFFSET_PE_AFTER_MZ = 0x200
 
 
-def find_embedded_pe(block_bytez: bytes, mz_xor: list[tuple[bytes, bytes, int]]) -> Iterator[tuple[int, int]]:
+def find_embedded_pe(
+    block_bytez: bytes, mz_xor: list[tuple[bytes, bytes, int]]
+) -> Iterator[tuple[int, int]]:
     """check segment for embedded PE
 
     adapted for Ghidra from:
@@ -52,7 +66,9 @@ def find_embedded_pe(block_bytez: bytes, mz_xor: list[tuple[bytes, bytes, int]])
             continue
 
         e_lfanew_bytes = block_bytez[e_lfanew : e_lfanew + 4]
-        newoff = struct.unpack("<I", capa.features.extractors.helpers.xor_static(e_lfanew_bytes, i))[0]
+        newoff = struct.unpack(
+            "<I", capa.features.extractors.helpers.xor_static(e_lfanew_bytes, i)
+        )[0]
 
         # assume XOR'd "PE" bytes exist within threshold
         if newoff > MAX_OFFSET_PE_AFTER_MZ:
@@ -80,11 +96,19 @@ def extract_file_embedded_pe() -> Iterator[tuple[Feature, Address]]:
         for i in range(256)
     ]
 
-    for block in capa.features.extractors.ghidra.helpers.get_current_program().getMemory().getBlocks():
-        if not all((block.isLoaded(), block.isInitialized(), "Headers" not in block.getName())):
+    for block in (
+        capa.features.extractors.ghidra.helpers.get_current_program()
+        .getMemory()
+        .getBlocks()
+    ):
+        if not all(
+            (block.isLoaded(), block.isInitialized(), "Headers" not in block.getName())
+        ):
             continue
 
-        for off, _ in find_embedded_pe(capa.features.extractors.ghidra.helpers.get_block_bytes(block), mz_xor):
+        for off, _ in find_embedded_pe(
+            capa.features.extractors.ghidra.helpers.get_block_bytes(block), mz_xor
+        ):
             # add offset back to block start
             ea_addr = block.getStart().add(off)
             ea = ea_addr.getOffset()
@@ -119,10 +143,18 @@ def extract_file_export_names() -> Iterator[tuple[Feature, Address]]:
                             libname = libname[:-4]
 
                         forwarded_name = f"{libname}.{ext_loc.getLabel()}"
-                        forwarded_name = capa.features.extractors.helpers.reformat_forwarded_export_name(forwarded_name)
+                        forwarded_name = capa.features.extractors.helpers.reformat_forwarded_export_name(
+                            forwarded_name
+                        )
 
-                        yield Export(forwarded_name), AbsoluteVirtualAddress(addr.getOffset())
-                        yield Characteristic("forwarded export"), AbsoluteVirtualAddress(addr.getOffset())
+                        yield (
+                            Export(forwarded_name),
+                            AbsoluteVirtualAddress(addr.getOffset()),
+                        )
+                        yield (
+                            Characteristic("forwarded export"),
+                            AbsoluteVirtualAddress(addr.getOffset()),
+                        )
                         is_forwarded = True
                         break
 
@@ -142,30 +174,56 @@ def extract_file_import_names() -> Iterator[tuple[Feature, Address]]:
      - importname
     """
 
-    for f in capa.features.extractors.ghidra.helpers.get_current_program().getFunctionManager().getExternalFunctions():
+    for f in (
+        capa.features.extractors.ghidra.helpers.get_current_program()
+        .getFunctionManager()
+        .getExternalFunctions()
+    ):
+        addr = None
         for r in f.getSymbol().getReferences():
             if r.getReferenceType().isData():
-                addr = r.getFromAddress().getOffset()  # gets pointer to fake external addr
+                addr = (
+                    r.getFromAddress().getOffset()
+                )  # gets pointer to fake external addr
+                break
 
-        fstr = f.toString().split("::")  # format: MODULE.dll::import / MODULE::Ordinal_*
+        if addr is None:
+            continue
+
+        fstr = f.toString().split(
+            "::"
+        )  # format: MODULE.dll::import / MODULE::Ordinal_*
         if "Ordinal_" in fstr[1]:
             fstr[1] = f"#{fstr[1].split('_')[1]}"
 
-        for name in capa.features.extractors.helpers.generate_symbols(fstr[0][:-4], fstr[1], include_dll=True):
+        for name in capa.features.extractors.helpers.generate_symbols(
+            fstr[0][:-4], fstr[1], include_dll=True
+        ):
             yield Import(name), AbsoluteVirtualAddress(addr)
 
 
 def extract_file_section_names() -> Iterator[tuple[Feature, Address]]:
     """extract section names"""
 
-    for block in capa.features.extractors.ghidra.helpers.get_current_program().getMemory().getBlocks():
-        yield Section(block.getName()), AbsoluteVirtualAddress(block.getStart().getOffset())
+    for block in (
+        capa.features.extractors.ghidra.helpers.get_current_program()
+        .getMemory()
+        .getBlocks()
+    ):
+        yield (
+            Section(block.getName()),
+            AbsoluteVirtualAddress(block.getStart().getOffset()),
+        )
 
 
 def extract_file_strings() -> Iterator[tuple[Feature, Address]]:
     """extract ASCII and UTF-16 LE strings"""
 
-    for block in capa.features.extractors.ghidra.helpers.get_current_program().getMemory().getBlocks():
+    for block in (
+        capa.features.extractors.ghidra.helpers.get_current_program()
+        .getMemory()
+        .getBlocks()
+    ):
         if not block.isInitialized():
             continue
 
@@ -185,9 +243,17 @@ def extract_file_function_names() -> Iterator[tuple[Feature, Address]]:
     extract the names of statically-linked library functions.
     """
 
-    for sym in capa.features.extractors.ghidra.helpers.get_current_program().getSymbolTable().getAllSymbols(True):
+    for sym in (
+        capa.features.extractors.ghidra.helpers.get_current_program()
+        .getSymbolTable()
+        .getAllSymbols(True)
+    ):
         # .isExternal() misses more than this config for the function symbols
-        if sym.getSymbolType() == SymbolType.FUNCTION and sym.getSource() == SourceType.ANALYSIS and sym.isGlobal():
+        if (
+            sym.getSymbolType() == SymbolType.FUNCTION
+            and sym.getSource() == SourceType.ANALYSIS
+            and sym.isGlobal()
+        ):
             name = sym.getName()  # starts to resolve names based on Ghidra's FidDB
             if name.startswith("FID_conflict:"):  # format: FID_conflict:<function-name>
                 name = name[13:]