+++ capa/features/extractors/ghidra/helpers.py
@@ -49,7 +49,9 @@ def ints_to_bytes(bytez: list[int]) -> bytes:
     return bytes([b & 0xFF for b in bytez])
 
 
-def find_byte_sequence(addr: "ghidra.program.model.address.Address", seq: bytes) -> Iterator[int]:
+def find_byte_sequence(
+    addr: "ghidra.program.model.address.Address", seq: bytes
+) -> Iterator[int]:
     """yield all ea of a given byte sequence
 
     args:
@@ -102,19 +104,28 @@ def get_function_symbols():
     yield from get_current_program().getFunctionManager().getFunctionsNoStubs(True)
 
 
-def get_function_blocks(fh: "capa.features.extractors.base_extractor.FunctionHandle") -> Iterator[BBHandle]:
+def get_function_blocks(
+    fh: "capa.features.extractors.base_extractor.FunctionHandle",
+) -> Iterator[BBHandle]:
     """
     yield the basic blocks of the function
     """
 
-    for block in SimpleBlockIterator(BasicBlockModel(get_current_program()), fh.inner.getBody(), get_monitor()):
-        yield BBHandle(address=AbsoluteVirtualAddress(block.getMinAddress().getOffset()), inner=block)
+    for block in SimpleBlockIterator(
+        BasicBlockModel(get_current_program()), fh.inner.getBody(), get_monitor()
+    ):
+        yield BBHandle(
+            address=AbsoluteVirtualAddress(block.getMinAddress().getOffset()),
+            inner=block,
+        )
 
 
 def get_insn_in_range(bbh: BBHandle) -> Iterator[InsnHandle]:
     """yield InshHandle for each insn in a given basicblock"""
     for insn in get_current_program().getListing().getInstructions(bbh.inner, True):
-        yield InsnHandle(address=AbsoluteVirtualAddress(insn.getAddress().getOffset()), inner=insn)
+        yield InsnHandle(
+            address=AbsoluteVirtualAddress(insn.getAddress().getOffset()), inner=insn
+        )
 
 
 def get_file_imports() -> dict[int, list[str]]:
@@ -123,13 +134,24 @@ def get_file_imports() -> dict[int, list[str]]:
     import_dict: dict[int, list[str]] = {}
 
     for f in get_current_program().getFunctionManager().getExternalFunctions():
+        addr = None
         for r in f.getSymbol().getReferences():
             if r.getReferenceType().isData():
-                addr = r.getFromAddress().getOffset()  # gets pointer to fake external addr
+                addr = (
+                    r.getFromAddress().getOffset()
+                )  # gets pointer to fake external addr
+                break
 
-        ex_loc = f.getExternalLocation().getAddress()  # map external locations as well (offset into module files)
+        if addr is None:
+            continue
 
-        fstr = f.toString().split("::")  # format: MODULE.dll::import / MODULE::Ordinal_* / <EXTERNAL>::import
+        ex_loc = (
+            f.getExternalLocation().getAddress()
+        )  # map external locations as well (offset into module files)
+
+        fstr = f.toString().split(
+            "::"
+        )  # format: MODULE.dll::import / MODULE::Ordinal_* / <EXTERNAL>::import
         if "Ordinal_" in fstr[1]:
             fstr[1] = f"#{fstr[1].split('_')[1]}"
 
@@ -162,7 +184,11 @@ def get_file_externs() -> dict[int, list[str]]:
 
     for sym in get_current_program().getSymbolTable().getAllSymbols(True):
         # .isExternal() misses more than this config for the function symbols
-        if sym.getSymbolType() == SymbolType.FUNCTION and sym.getSource() == SourceType.ANALYSIS and sym.isGlobal():
+        if (
+            sym.getSymbolType() == SymbolType.FUNCTION
+            and sym.getSource() == SourceType.ANALYSIS
+            and sym.isGlobal()
+        ):
             name = sym.getName()  # starts to resolve names based on Ghidra's FidDB
             if name.startswith("FID_conflict:"):  # format: FID_conflict:<function-name>
                 name = name[13:]
@@ -172,7 +198,9 @@ def get_file_externs() -> dict[int, list[str]]:
                 # extract features for both the mangled and un-mangled representations.
                 # e.g. `_fwrite` -> `fwrite`
                 # see: https://stackoverflow.com/a/2628384/87207
-                extern_dict.setdefault(sym.getAddress().getOffset(), []).append(name[1:])
+                extern_dict.setdefault(sym.getAddress().getOffset(), []).append(
+                    name[1:]
+                )
 
     return extern_dict
 
@@ -201,7 +229,9 @@ def map_fake_import_addrs() -> dict[int, list[int]]:
     for f in get_current_program().getFunctionManager().getExternalFunctions():
         for r in f.getSymbol().getReferences():
             if r.getReferenceType().isData():
-                fake_dict.setdefault(f.getEntryPoint().getOffset(), []).append(r.getFromAddress().getOffset())
+                fake_dict.setdefault(f.getEntryPoint().getOffset(), []).append(
+                    r.getFromAddress().getOffset()
+                )
 
     return fake_dict
 
@@ -230,13 +260,18 @@ def check_addr_for_api(
 
 
 def is_call_or_jmp(insn: "ghidra.program.database.code.InstructionDB") -> bool:
-    return any(mnem in insn.getMnemonicString() for mnem in ["CALL", "J"])  # JMP, JNE, JNZ, etc
+    return any(
+        mnem in insn.getMnemonicString() for mnem in ["CALL", "J"]
+    )  # JMP, JNE, JNZ, etc
 
 
 def is_sp_modified(insn: "ghidra.program.database.code.InstructionDB") -> bool:
     for i in range(insn.getNumOperands()):
         if insn.getOperandType(i) == OperandType.REGISTER:
-            return "SP" in insn.getRegister(i).getName() and insn.getOperandRefType(i).isWrite()
+            return (
+                "SP" in insn.getRegister(i).getName()
+                and insn.getOperandRefType(i).isWrite()
+            )
     return False